+++ src/main/java/com/dev/trabProjarq/dominio/services/IOcupacaoAeroviaRep.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface IOcupacaoAeroviaRep {
-    List<OcupacaoAerovia> findOcupadasSlots(int aeroviaId, LocalDate data, List<Float> slotsHorarios);
+    List<OcupacaoAerovia> findOcupadasSlots(int aeroviaId, LocalDate data, List<Integer> slotsHorarios);
     List<OcupacaoAerovia> findAllOcupacaoAeroviasByData(int aeroviaId, LocalDate data);
 
     OcupacaoAerovia ocupa(OcupacaoAerovia ocupacaoAerovia);
